Renamed with-all-extensions.js -> index.js
